LayoutTests/imported/w3c/web-platform-tests/loading/early-hints/referrer-policy-same-origin.h2.window-expected.txt

CONSOLE MESSAGE: <link rel=preload> has an invalid `href` value

FAIL Referrer policy: same-origin assert_true: expected true got false
